Add start/stop button
This gives us a simple start/stop button that
mostly works. When the computation gets really
slow (fib for large n) it's not always as
responsive as I'd like. I'm not sure if we can
actually fix that as I don't think there's any
way to interrupt a computation of `(fib n)`
while it's in progress.

I do think I want to explore having a proper
events queue, though, and inserting start/stop
events into that queue to see if that improves
things.