Get FontAwesome buttons working
Instead of "Start" and "Stop" this now has
play and pause icons with colors. Cryogen
was pulling an old version (4.2) of an old version of FontAwesome 
(they're on 5.x now).
I updated it to 4.7.0 (last of the 4.x) so
I could get the circled versions of play and
pause; we should probably upgrade all the way
to 5.x unless that will break something.