Switch to input-output channels
This changes things so we have input/output
channels for the `fib` function, which allows
us to remove the "hacky" `timeout` calls. (The
fact that the delay to put in those was quite 
arbitrary was definitley a smell.) 

I think this has actually improved the
responsiveness of the buttons, which is cool.

Now I think the trick is to sort out a more
coherent "event queue" so we have proper
`:running`, `:paused`, and `:stopped` states.